Refactor: remove unused imports + cleanup file structure
